+++ src/BodyParser.php
@@ -78,7 +78,7 @@ class BodyParser
 
     /**
      * Extracs links from body
-     * @return array
+     * @return array Array of links
      */
     public function extractLinks(): array
     {