+++ src/Exception/PropertyNotExistsException.php
@@ -19,7 +19,7 @@ namespace Tools\Exception;
 use Exception;
 
 /**
- * "Property does not exist" exception
+ * "Property does not exist" exception.
  */
 class PropertyNotExistsException extends Exception
 {