+++ src/TestSuite/TestTrait.php
@@ -21,7 +21,7 @@ use Exception;
 use Throwable;
 
 /**
- * A trait that provides some assertion methods
+ * A trait that provides some assertion methods.
  * @method void assertIsArray($var, $message = '') Asserts that `$var` is an array
  * @method void assertIsBool($var, $message = '') Asserts that `$var` is a boolean
  * @method void assertIsCallable($var, $message = '') Asserts that `$var` is a callable